Do not accept user creation with existent email
Uniqueness email validation is done at the database layer, so no race
condition will happen by two different application processes which are
modifying the database simultanously.

The UniquenessConstraint exception is handled in the controller and a
ValidationErrorException is raised. This ensures the same behaviour as
it would have been implemented as an uniqueness validation handler by
Hanami::Validator.