Generates user token after its creation
I use the SecureRandom module for generating a random hex string with 32
characters. I did not override the User#initialize, because the entity
should not be enforced to assign a token. Maybe there are users which do
not need a token in the future. Another reason is, that it is better
testable.